+++ rust/src/page_mapper.rs
@@ -1,4 +1,8 @@
-use crate::utils::{invalid_data, read_skipping_ws};
+//  Copyright (C) 2025 Intel Corporation
+//
+//  SPDX-License-Identifier: MIT
+
+use crate::utils::{invalid_data, read_skipping_ws, skip_serde_json_value};
 use std::io::{Error, Read, Seek};
 
 pub trait ParsedJsonSection: Sized {
@@ -9,7 +13,7 @@ pub trait JsonPageMapper<T>: Sized
 where
     T: ParsedJsonSection,
 {
-    fn parse_json(mut reader: impl Read + Seek) -> Result<Vec<Box<T>>, Error> {
+    fn parse_json(mut reader: impl Read + Seek, strict: bool) -> Result<Vec<Box<T>>, Error> {
         let mut brace_level = 0;
         let mut json_sections = Vec::new();
 
@@ -26,8 +30,21 @@ where
                     }
                     match String::from_utf8(buf_key.clone()) {
                         Ok(key) => {
-                            let section = T::parse(key, &mut reader)?;
-                            json_sections.push(section);
+                            match T::parse(key.clone(), &mut reader) {
+                                Ok(section) => {
+                                    json_sections.push(section);
+                                }
+                                Err(e) => {
+                                    if strict {
+                                        return Err(e);
+                                    } else {
+                                        // Skip unknown section
+                                        eprintln!("Skipping unknown section key: \"{}\"", key);
+                                        read_skipping_ws(&mut reader)?;
+                                        let _ = skip_serde_json_value(&mut reader);
+                                    }
+                                }
+                            }
                         }
                         Err(e) => {
                             let cur_pos = reader.stream_position()?;